split out CalendarEventDetailsForm tests
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/367745
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
