+++ ui/features/calendar/react/__tests__/CalendarEventDetailsForm2.test.jsx
@@ -110,229 +110,6 @@ describe('CalendarEventDetailsForm', () => {
     jest.clearAllMocks()
   })
 
-  it('creates a new event', async () => {
-    defaultProps.event.isNewEvent = () => true
-    const component = render(<CalendarEventDetailsForm {...defaultProps} />)
-    expect(defaultProps.setSetContextCB).toHaveBeenCalled()
-    expect(defaultProps.contextChangeCB).toHaveBeenCalled()
-    component.getByText('Submit').click()
-
-    await waitFor(() => expect(defaultProps.closeCB).toHaveBeenCalled())
-
-    // event.possibleContexts() is only called when a the event is new.
-    expect(defaultProps.event.possibleContexts).toHaveBeenCalled()
-    defaultProps.event.isNewEvent = () => false
-  })
-
-  it('renders main elements and updates an event with valid parameters (flaky)', async () => {
-    const component = render(<CalendarEventDetailsForm {...defaultProps} />)
-
-    changeValue(component, 'edit-calendar-event-form-title', 'Class Party')
-    changeValue(component, 'edit-calendar-event-form-location', 'The Zoo')
-    changeValue(component, 'edit-calendar-event-form-date', '2022-07-23T00:00:00.000Z')
-    setTime(component, 'event-form-start-time', '2:00 AM')
-    setTime(component, 'event-form-end-time', '3:00 PM')
-    component.getByText('Calendar').click()
-    component.getByText('Geometry').click()
-    expect(component.getByText('More Options')).toBeInTheDocument()
-    component.getByText('Submit').click()
-    await waitFor(() => expect(defaultProps.closeCB).toHaveBeenCalled())
-    expect(defaultProps.event.save).toHaveBeenCalledWith(
-      expect.objectContaining({
-        'calendar_event[title]': 'Class Party',
-        'calendar_event[start_at]': '2022-07-23T02:00:00.000Z',
-        'calendar_event[end_at]': '2022-07-23T15:00:00.000Z',
-        'calendar_event[location_name]': 'The Zoo',
-        'calendar_event[web_conference]': '',
-        'calendar_event[context_code]': 'course_1',
-        'calendar_event[important_dates]': false,
-        'calendar_event[blackout_date]': false,
-      }),
-      expect.any(Function),
-      expect.any(Function),
-    )
-  })
-
-  it('shows UpdateCalendarEventsDialog when saving a recurring event', async () => {
-    const props = eventFormProps()
-    props.event.calendarEvent = {
-      ...props.event.calendarEvent,
-      series_uuid: '123',
-    }
-    props.event.object = {rrule: 'FREQ=DAILY;INTERVAL=1;COUNT=3'}
-
-    const component = render(<CalendarEventDetailsForm {...props} />)
-    component.getByText('Submit').click()
-
-    expect(UpdateCalendarEventDialogModule.renderUpdateCalendarEventDialog).toHaveBeenCalled()
-    await waitFor(() =>
-      expect(props.event.save).toHaveBeenCalledWith(
-        expect.objectContaining({which: 'all'}),
-        expect.any(Function),
-        expect.any(Function),
-      ),
-    )
-  })
-
-  it('does not show UpdateCalendarEventsDialog when saving a single event', async () => {
-    const props = eventFormProps()
-    props.event.calendarEvent = {
-      ...props.event.calendarEvent,
-      series_uuid: null,
-    }
-    props.event.object = {rrule: null}
-
-    const component = render(<CalendarEventDetailsForm {...props} />)
-    component.getByText('Submit').click()
-
-    expect(UpdateCalendarEventDialogModule.renderUpdateCalendarEventDialog).not.toHaveBeenCalled()
-    await waitFor(() => expect(props.event.save).toHaveBeenCalled())
-  })
-
-  it('does not show UpdateCalendarEventsDialog when changing series to a single event', async () => {
-    const props = eventFormProps()
-    props.event.calendarEvent = {
-      ...props.event.calendarEvent,
-      series_uuid: '123',
-    }
-    props.event.object = {rrule: 'FREQ=DAILY;INTERVAL=1;COUNT=3'}
-
-    const component = render(<CalendarEventDetailsForm {...props} />)
-    component.getByText('Frequency').click() // open the dropdown
-    component.getByText('Does not repeat').click() // select the option
-    component.getByText('Submit').click()
-
-    expect(UpdateCalendarEventDialogModule.renderUpdateCalendarEventDialog).not.toHaveBeenCalled()
-    await waitFor(() => expect(props.event.save).toHaveBeenCalled())
-  })
-
-  it('can change the date multiple times', async () => {
-    const component = render(<CalendarEventDetailsForm {...defaultProps} />)
-
-    let date = changeValue(component, 'edit-calendar-event-form-date', '2022-07-03T00:00:00.000Z')
-    expect(date.value).toBe('Sun, Jul 3, 2022')
-    date = changeValue(component, 'edit-calendar-event-form-date', '2022-07-14T00:00:00.000Z')
-    expect(date.value).toBe('Thu, Jul 14, 2022')
-    date = changeValue(component, 'edit-calendar-event-form-date', '2022-07-23T00:00:00.000Z')
-    expect(date.value).toBe('Sat, Jul 23, 2022')
-    component.getByText('Submit').click()
-
-    await waitFor(() =>
-      expect(defaultProps.event.save).toHaveBeenCalledWith(
-        expect.objectContaining({
-          'calendar_event[start_at]': '2022-07-23T00:00:00.000Z',
-        }),
-        expect.any(Function),
-        expect.any(Function),
-      ),
-    )
-  })
-
-  it('shows the date for the France locale', async () => {
-    const old_locale = window.ENV.LOCALE
-    window.ENV.LOCALE = 'fr'
-
-    const props = {...defaultProps}
-    props.event = commonEventFactory(null, [userContext, courseContext])
-    const d = moment('2023-08-28') // a monday
-    props.event.date = d.toDate()
-    props.event.startDate = jest.fn(() => moment(props.event.date))
-
-    const {getByTestId} = render(<CalendarEventDetailsForm {...props} />)
-    const beginning_date = getByTestId('edit-calendar-event-form-date')
-
-    expect(beginning_date.value).toBe('lun. 28 août 2023')
-    window.ENV.LOCALE = old_locale
-  })
-
-  // LF-630 (08/23/2023)
-  it.skip('can keep the same date when the date input is clicked and blurred', async () => {
-    const component = render(<CalendarEventDetailsForm {...defaultProps} />)
-
-    const date = changeValue(component, 'edit-calendar-event-form-date', '2022-07-14T00:00:00.000Z')
-    expect(date.value).toBe('Thu, Jul 14, 2022')
-
-    for (let i = 0; i < 30; i++) {
-      act(() => date.click())
-      act(() => date.blur())
-    }
-
-    component.getByText('Submit').click()
-
-    await waitFor(() =>
-      expect(defaultProps.event.save).toHaveBeenCalledWith(
-        expect.objectContaining({
-          'calendar_event[start_at]': '2022-07-14T00:00:00.000Z',
-        }),
-        expect.any(Function),
-        expect.any(Function),
-      ),
-    )
-  })
-
-  it('can change the date in Denver at 12:00 AM', async () => {
-    testTimezone('America/Denver', '2022-07-14T06:00:00.000Z', '2022-07-14T06:00:00.000Z')
-  })
-
-  it('can change the date in Denver at 11:30 PM', async () => {
-    testTimezone(
-      'America/Denver',
-      '2022-07-14T06:00:00.000Z',
-      '2022-07-15T05:30:00.000Z',
-      '11:30 PM',
-    )
-  })
-
-  it('can change the date in Shanghai at 12:00 AM', async () => {
-    testTimezone('Asia/Shanghai', '2022-07-13T16:00:00.000Z', '2022-07-13T16:00:00.000Z')
-  })
-
-  it('can change the date in Shanghai at 11:30 PM', async () => {
-    testTimezone(
-      'Asia/Shanghai',
-      '2022-07-13T16:00:00.000Z',
-      '2022-07-14T15:30:00.000Z',
-      '11:30 PM',
-    )
-  })
-
-  it('can change the date in Adelaide at 12:00 AM', async () => {
-    testTimezone('Australia/Adelaide', '2022-07-13T14:30:00.000Z', '2022-07-13T14:30:00.000Z')
-  })
-
-  it('can change the date in Adelaide at 11:30 PM', async () => {
-    testTimezone(
-      'Australia/Adelaide',
-      '2022-07-13T14:30:00.000Z',
-      '2022-07-14T14:00:00.000Z',
-      '11:30 PM',
-    )
-  })
-
-  it('can change the date in Tokyo at 12:00 AM', async () => {
-    testTimezone('Asia/Tokyo', '2022-07-13T15:00:00.000Z', '2022-07-13T15:00:00.000Z')
-  })
-
-  it('can change the date in Tokyo at 11:30 PM', async () => {
-    testTimezone('Asia/Tokyo', '2022-07-13T15:00:00.000Z', '2022-07-14T14:30:00.000Z', '11:30 PM')
-  })
-
-  it('can change the date in the UK at 12:00 AM', async () => {
-    testTimezone('Etc/UTC', '2022-07-14T00:00:00.000Z', '2022-07-14T00:00:00.000Z')
-  })
-
-  it('can change the date in the UK at 11:30 PM', async () => {
-    testTimezone('Etc/UTC', '2022-07-14T00:00:00.000Z', '2022-07-14T23:30:00.000Z', '11:30 PM')
-  })
-
-  it('can change the date in eastern Brazil at 12:00 AM', async () => {
-    testTimezone('Brazil/East', '2022-07-14T03:00:00.000Z', '2022-07-14T03:00:00.000Z')
-  })
-
-  it('can change the date in eastern Brazil at 11:30 PM', async () => {
-    testTimezone('Brazil/East', '2022-07-14T03:00:00.000Z', '2022-07-15T02:30:00.000Z', '11:30 PM')
-  })
-
   it('cannot have end time before start time', async () => {
     const component = render(<CalendarEventDetailsForm {...defaultProps} />)
